% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensator.R
\name{compensator}
\alias{compensator}
\alias{compensator.default}
\alias{compensator.mmpp}
\alias{compensator.hp}
\alias{compensator.mmhp}
\alias{compensator.hpp}
\title{Compensators for mmhp, hp and hpp models}
\usage{
compensator(object, events)

\method{compensator}{default}(object, events)

\method{compensator}{mmpp}(object, events)

\method{compensator}{hp}(object, events)

\method{compensator}{mmhp}(object, events)

\method{compensator}{hpp}(object, events)
}
\arguments{
\item{object}{a social network model}

\item{events}{event times, which can have first value as 0}
}
\value{
compensator vector of rescaled interevent times
}
\description{
Compensators for mmhp, hp and hpp models
}
\examples{
hpp_obj <- pp_hpp(lambda = 1)
events <- pp_simulate(hpp_obj, end=10)
comp <- compensator(hpp_obj, events)
}
