% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHPIntensity.R
\name{drawHPIntensity}
\alias{drawHPIntensity}
\title{Draw the intensity of Hawkes Process}
\usage{
drawHPIntensity(hp, start, end, history = 0, t, color = 1, i = 1, add = FALSE)
}
\arguments{
\item{hp}{object parameters for Hawkes process}

\item{start}{the start time of current state}

\item{end}{the end time of current state}

\item{history}{the past event times}

\item{t}{the event times happened in this state}

\item{color}{A specification for the default plotting color.}

\item{i}{state number this corresponds to a state jump directly before
which is only important when using mmhp}

\item{add}{whether to add the hawkes intensity to an existing plot}
}
\description{
Draw the intensity of Hawkes Process, a helper function for 'drawUniMMHPIntensity'
while also available independently
}
\examples{
hp_obj <- hp(lambda0 = 0.1,alpha = 0.45,beta = 0.5)
sims <- simulatehp(hp_obj,start = 0, end = 100, history = 0)
events=sims$t
drawHPIntensity(hp_obj, start=0, end=max(events), history=0, t=events)
}
