% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHPPIntensity.R
\name{drawHPPIntensity}
\alias{drawHPPIntensity}
\title{Draw Intensity of Homogeneous Poisson Process}
\usage{
drawHPPIntensity(
  hpp,
  events,
  color = "red",
  start = 0,
  end = max(events),
  fit = FALSE,
  int_title = "Intensity homogeneous Poisson Process"
)
}
\arguments{
\item{hpp}{object for homogeneous poisson process}

\item{events}{event times input}

\item{color}{a specification for the default plotting color.}

\item{start}{start of events}

\item{end}{end of events}

\item{fit}{boolean indicating whether to fit hpp object}

\item{int_title}{the plot title}
}
\description{
Draw the intensity for homogeneous poisson process events
}
\examples{
\dontrun{
pois_y <- hpp(lambda = 1)
drawHPPIntensity(pois_y, events=simulatehpp(pois_y, end=10), color = "red")
}
}
