% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetworkhp.R
\name{fit.network.hp}
\alias{fit.network.hp}
\title{Fit multiple hawkes processes to a network using negloglik and optim functions}
\usage{
fit.network.hp(vec, t, termination)
}
\arguments{
\item{vec}{vector containing initial values for the object parameters (lambda0,alpha,beta) to be optimized.}

\item{t}{network describing in a datatable containing the sender, receiver and event times.}

\item{termination}{the end time of event times.}
}
\value{
a list of hp object indicating the maximum likelihood parameter values (lambda0,alpha,beta) for hawkes process plus the sender and receiver information.
}
\description{
Compute the maximum likelihood parameter values for hawkes process.
}
