% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetworkhpp.R
\name{fit.network.hpp}
\alias{fit.network.hpp}
\title{Fit multiple poisson processes to a network}
\usage{
fit.network.hpp(t)
}
\arguments{
\item{t}{network describing in a datatable containing the sender, receiver and event times.}
}
\value{
a list of hpp object indicating the maximum likelihood parameter values (lambda0) for poisson process plus the sender and receiver information.
}
\description{
Fit multiple poisson processes to a network
}
