% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSingleNetworkhpp.R
\name{fit.single.network.hpp}
\alias{fit.single.network.hpp}
\title{Fit single poisson process to a network}
\usage{
fit.single.network.hpp(t)
}
\arguments{
\item{t}{network describing in a datatable containing the sender, receiver and event times.}
}
\value{
a single value indicating lambda for the single poisson process fitting to the whole network
}
\description{
Fit single poisson process to a network
}
