% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hppmle.R
\name{hppmle}
\alias{hppmle}
\title{Fit a homogeneous poisson process}
\usage{
hppmle(t)
}
\arguments{
\item{t}{vector containing the event times.}
}
\value{
a scalar indicating the maximum likelihood estimator
}
\description{
Compute maximum likelihood estimator for the hpp model with time events
}
