% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\alias{intensity.default}
\alias{intensity.mmhp}
\alias{intensity.hp}
\alias{intensity.mmpp}
\title{Compute the intensity of social network model}
\usage{
intensity(object, event, method = "function")

\method{intensity}{default}(object, event, method = "function")

\method{intensity}{mmhp}(object, event, method = "function")

\method{intensity}{hp}(object, event, method = "function")

\method{intensity}{mmpp}(object, event, method = "function")
}
\arguments{
\item{object}{an object of MMHP/HP/MMPP
For example, MMHP object should includ its state, state_time, tau, lambda0, lambda1, beta and alpha.}

\item{event}{the observed/simulated events}

\item{method}{the method used to calculate intensity.
The candidates are: `function`, `numeric`, and `atevent`, default to `function`.}
}
\value{
The intensity function of MMHP
}
\description{
Take an object of MMHP/HP/MMPP and generate its intensity function accordingly
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- simulatemmhp(x)
z <- intensity(x, y)
}
