% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksplot.R
\name{ksplot}
\alias{ksplot}
\title{KS plot to visualize the empirical and theoretical cdf curve}
\usage{
ksplot(r, ...)
}
\arguments{
\item{r}{rescaled-inter-event-times}

\item{...}{other arguments for plots}
}
\description{
Plot empirical cdf plot for rescaled-inter-event-times and
 exponential cdf as a reference curve
}
