% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpp.R
\name{mmpp}
\alias{mmpp}
\title{Create a Markov-modulated Poisson Process(MMPP) model}
\usage{
mmpp(lambda0, c, q1, q2, tau = NULL)
}
\arguments{
\item{lambda0}{parameters for Poisson process.}

\item{c}{the proportion of intensity 1 over intensity 2}

\item{q1}{transition probability}

\item{q2}{transition probability}

\item{tau}{vector containing the event times. Note that the first event is at time zero. Alternatively, tau could be specified as NULL, meaning that the data will be added later (e.g. simulated).}
}
\value{
mmpp object
}
\description{
Create a Markov-modulated Poisson Process(MMPP) model according to the given parameters: lambda0, c, q1, q2 and event times.
If event time tau is missing, then it means that data will be added later(e.g. simulated)
}
