% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglik.R
\name{negloglik}
\alias{negloglik}
\alias{negloglik.default}
\alias{negloglik.hp}
\alias{negloglik.mmhp}
\title{Compute negative log likelihood for social network models}
\usage{
negloglik(object, t, termination)

\method{negloglik}{default}(object, t, termination)

\method{negloglik}{hp}(object, t, termination)

\method{negloglik}{mmhp}(object, t, termination)
}
\arguments{
\item{object}{social network model contating the parameters}

\item{t}{vector containing the event times.}

\item{termination}{the end time of event times}
}
\value{
a scalar indicating the negative log likelihood
}
\description{
Compute negative log likelihood for social network models with model specified time events or simulated time events
}
