% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithp.R
\name{negloglik_hp_new}
\alias{negloglik_hp_new}
\title{Fit a hawkes process using negloglik and optim functions}
\usage{
negloglik_hp_new(vec, t, termination)
}
\arguments{
\item{vec}{vector containing initial values for the object parameters (lambda0,alpha,beta) to be optimized.}

\item{t}{vector containing event times.}

\item{termination}{the end time of event times.}
}
\value{
a hp object indicating the maximum likelihood parameter values (lambda0,alpha,beta) for hawkes process.
}
\description{
Compute the maximum likelihood parameter values for hawkes process.
}
