% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_hpp_network_pearson_residual.R
\name{network_pearsonresidual.hpp}
\alias{network_pearsonresidual.hpp}
\title{print a pearson residual plot for fitting single poisson process to a network}
\usage{
network_pearsonresidual.hpp(
  t,
  n_node,
  lambda0,
  termination,
  start_time,
  end_time
)
}
\arguments{
\item{t}{network describing in a datatable containing the sender, receiver and event times.}

\item{n_node}{number of nodes in the network}

\item{lambda0}{the common hp parameter lambdaa across all the pairs}

\item{start_time}{the starting time of the events}

\item{end_time}{the ending time of the events}
}
\value{
a pearson residual plot for fitting a single poisson process to a network
}
\description{
print a pearson residual plot for fitting single poisson process to a network
}
