% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_hpp_network_raw_residual.R
\name{network_rawresidual.hpp}
\alias{network_rawresidual.hpp}
\title{print a raw raw plot for fitting single poisson process to a network}
\usage{
network_rawresidual.hpp(t, n_node, lambda0, termination, start_time, end_time)
}
\arguments{
\item{t}{network describing in a datatable containing the sender, receiver and event times.}

\item{n_node}{number of nodes in the network}

\item{lambda0}{the common hp parameter lambdaa across all the pairs}

\item{start_time}{the starting time of the events}

\item{end_time}{the ending time of the events}
}
\value{
a raw residual plot for fitting a single poisson process to a network
}
\description{
print a raw raw plot for fitting single poisson process to a network
}
