% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkRawresidual.R
\name{networkRawresidual}
\alias{networkRawresidual}
\title{Compute raw residuals for social network models}
\usage{
networkRawresidual(object, t, termination, time.vec, latent.vec)
}
\arguments{
\item{object}{social network model contating the parameters}

\item{t}{a dataframe containg three columns:send, rec and event happening time}

\item{termination}{termination time}

\item{time.vec}{time segment to calculate the intensity for `numeric` method}

\item{latent.vec}{the probability of the latent space being in the active state}
}
\value{
a dataframe of three columns: Send, Rec and the corresponsding Pearson residual
}
\description{
Compute network based raw residuals for social network models with model specified time events or simulated time events
}
