% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothpp.R
\name{plothpp}
\alias{plothpp}
\title{Draw Homogeneous Poisson Process}
\usage{
plothpp(events, color)
}
\arguments{
\item{events}{homogeneous poisson process events}

\item{color}{a specification for the default plotting color.}
}
\description{
Draw the event time counts for homogeneous poisson process
}
