% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothpp_step.R
\name{plothpp_step}
\alias{plothpp_step}
\title{Draw Homogeneous Poisson Process}
\usage{
plothpp_step(events, color = "red")
}
\arguments{
\item{color}{A specification for the default plotting color.}

\item{object}{parameters for homogeneous poisson process}
}
\description{
Draw the event time counts for homogeneous poisson process
}
