% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_compensator.R
\name{pp_compensator}
\alias{pp_compensator}
\alias{pp_compensator.default}
\alias{pp_compensator.mmpp}
\alias{pp_compensator.hp}
\alias{pp_compensator.mmhp}
\alias{pp_compensator.hpp}
\title{Compensators for point processes}
\usage{
pp_compensator(object, events)

\method{pp_compensator}{default}(object, events)

\method{pp_compensator}{mmpp}(object, events)

\method{pp_compensator}{hp}(object, events)

\method{pp_compensator}{mmhp}(object, events)

\method{pp_compensator}{hpp}(object, events)
}
\arguments{
\item{object}{a point process model}

\item{events}{event times, which can have first value as 0}
}
\value{
compensator vector of rescaled interevent times
}
\description{
Computes the compensator for included point processes
}
\examples{
hpp_obj <- pp_hpp(lambda = 1)
events <- pp_simulate(hpp_obj, end=10)
comp <- pp_compensator(hpp_obj, events)
}
