% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_hp.R
\name{pp_hp}
\alias{pp_hp}
\title{Create a Hawkes Process object}
\usage{
pp_hp(lambda0, alpha, beta, events = NULL)
}
\arguments{
\item{lambda0}{initial intensity at the start time}

\item{alpha}{jump size in increase of intensity}

\item{beta}{exponential decay of intensity}

\item{events}{vector containing the event times. 
Note that the first event is at time zero. 
Alternatively, events could be specified as NULL, 
meaning that the data will be added later (e.g. simulated).}
}
\value{
hp object
}
\description{
Create a Hawkes Process with an exponential
kernel according to the given parameters:
lambda0, alpha, beta and events.
If events are missing, then it means that data will be
added later(simulated from this process)
}
\examples{
pp_hp(lambda0 = 0.1,alpha = 0.45,beta = 0.5)
}
