% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_ksplot.R
\name{pp_ksplot}
\alias{pp_ksplot}
\title{KS plot of empirical and theoretical cdf curve of fitted point process}
\usage{
pp_ksplot(r, ...)
}
\arguments{
\item{r}{rescaled-inter-event-times}

\item{...}{other arguments for plots}
}
\value{
no return value, KS plot for rescaled-inter-event-times and exponential cdf curve
}
\description{
Plot empirical cdf plot for rescaled-inter-event-times and
 exponential cdf as a reference curve
}
