% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_qqexp.R
\name{pp_qqexp}
\alias{pp_qqexp}
\title{Plot QQ-plot for rescaled-inter-event-times of fitted point process}
\usage{
pp_qqexp(r, ...)
}
\arguments{
\item{r}{rescaled-inter-event-times}

\item{...}{other arguments for plots}
}
\value{
no return value, quantile-quantile plot for rescaled-inter-event-times
}
\description{
Generate Quantile-quantile plot for rescaled-inter-event-times,
which are independently
and identically distributed as exponential random variables with rate 1
under the true point process.
}
