% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_simulate.R
\name{pp_simulate.default}
\alias{pp_simulate.default}
\title{@rdname pp_simulate
 @export}
\usage{
\method{pp_simulate}{default}(object, events, start, end = 1, n = NULL)
}
\description{
@rdname pp_simulate
 @export
}
