% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqexp.R
\name{qqexp}
\alias{qqexp}
\title{Plot Quantile-quantile plot for rescaled-inter-event-times}
\usage{
qqexp(r, ...)
}
\arguments{
\item{r}{rescaled-inter-event-times}

\item{...}{other arguments for plots}
}
\description{
Generate Quantile-quantile plot for rescaled-inter-event-times,
 which are supposedly to be independently 
 and identically distributed as exponential random vairbales with rate 1.
}
