% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawresidual.R
\name{rawresidual}
\alias{rawresidual}
\alias{rawresidual.default}
\alias{rawresidual.hp}
\alias{rawresidual.mmhp}
\title{Compute raw residuals for social network models}
\usage{
rawresidual(object, t, termination, time.vec, latent.vec)

\method{rawresidual}{default}(object, t, termination, time.vec, latent.vec)

\method{rawresidual}{hp}(object, t, termination, time.vec, latent.vec)

\method{rawresidual}{mmhp}(object, t, termination, time.vec, latent.vec)
}
\arguments{
\item{object}{social network model contating the parameters}

\item{t}{vector of event happening time}

\item{termination}{termination time}

\item{time.vec}{time segment to calculate the intensity for `numeric` method}

\item{latent.vec}{the probability of the latent space being in the active state}
}
\value{
the raw residual
}
\description{
Compute raw residuals for social network models with model specified time events or simulated time events
}
