% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateNetworkhp.R
\name{simulateNetworkhp}
\alias{simulateNetworkhp}
\title{Simulate network based on Hawkes process during active state}
\usage{
simulateNetworkhp(
  lambda_matrix,
  alpha_matrix,
  beta_matrix,
  Z,
  start,
  horizon,
  history
)
}
\arguments{
\item{lambda_matrix}{parameters for Hawkes process.}

\item{alpha_matrix}{parameters for Hawkes process.}

\item{beta_matrix}{parameters for Hawkes process.}

\item{Z}{vector indicaating which group each node belongs to.}

\item{start}{start time of the Hawkes process.}

\item{horizon}{end time of the Hawkes process.}

\item{history}{the past event times.}
}
\value{
simulated network Hawkes Process in dataframe format
}
\description{
Simulate network based on Hawkes process during active state (including all the histoty), a helper function for "simulationmmhp"
}
