% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateNetworkhpp.R
\name{simulateNetworkhpp}
\alias{simulateNetworkhpp}
\title{Simulate network based on poisson process}
\usage{
simulateNetworkhpp(lambda_matrix, Z, start, horizon, history)
}
\arguments{
\item{lambda_matrix}{parameters for poisson process.}

\item{Z}{vector indicaating which group each node belongs to.}

\item{start}{start time of the poisson process.}

\item{horizon}{end time of the poisson process.}

\item{history}{the past event times.}
}
\value{
simulated network poisson Process in dataframe format
}
\description{
Simulate network based on poisson process (including all the histoty)
}
