% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensator.R
\name{compensator}
\alias{compensator}
\alias{compensator.default}
\alias{compensator.mmpp}
\alias{compensator.hp}
\alias{compensator.mmhp}
\alias{compensator.hpp}
\title{Compensators for MMHP, HP HPP, and MMPP models}
\usage{
compensator(object, events, pzt)

\method{compensator}{default}(object, events, pzt)

\method{compensator}{mmpp}(object, events, pzt)

\method{compensator}{hp}(object, events, pzt = NULL)

\method{compensator}{mmhp}(object, events, pzt)

\method{compensator}{hpp}(object, events, pzt = NULL)
}
\arguments{
\item{object}{a social network model}

\item{events}{event times}

\item{pzt}{probability}
}
\value{
vector of compensator
}
\description{
Compensators for MMHP, HP HPP, and MMPP models
}
