% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagpp.R
\name{diagpp}
\alias{diagpp}
\alias{diagpp.default}
\alias{diagpp.hp}
\alias{diagpp.mmhp}
\alias{diagpp.hpp}
\title{Generate diagnostic tools for different point process models, 
including quantile-quantile plot, ks plot, 
raw residual and pearson residual.}
\usage{
diagpp(object, events)

\method{diagpp}{default}(object, events)

\method{diagpp}{hp}(object, events)

\method{diagpp}{mmhp}(object, events)

\method{diagpp}{hpp}(object, events)
}
\arguments{
\item{object}{a point process model}

\item{events}{event times}
}
\value{
Display QQ-plot and KS plot, along with printing Pearson
 and raw residuals.
}
\description{
Generate diagnostic tools for different point process models, 
including quantile-quantile plot, ks plot, 
raw residual and pearson residual.
}
\examples{
\dontrun{
hpp_obj <- pp_hpp(lambda = 1)
events <- pp_simulate(hpp_obj, end=10)
diagpp(hpp_obj, events)
}
}
