% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagpp.R
\name{diagpp}
\alias{diagpp}
\alias{diagpp.default}
\alias{diagpp.hp}
\alias{diagpp.mmhp}
\alias{diagpp.hpp}
\title{Generate diagnostic tools for different models, 
including quantile-quantile plot, ks plot, 
raw residual and pearson residual.}
\usage{
diagpp(object, events, pzt = NULL)

\method{diagpp}{default}(object, events, pzt = NULL)

\method{diagpp}{hp}(object, events, pzt = NULL)

\method{diagpp}{mmhp}(object, events, pzt = 0.5)

\method{diagpp}{hpp}(object, events, pzt = NULL)
}
\arguments{
\item{object}{a social network model}

\item{events}{event times}

\item{pzt}{probability for producing compensator (mmhp)}
}
\value{
print qq plot and ks plot, and print out pearson and raw residuals.
}
\description{
Generate diagnostic tools for different models, 
including quantile-quantile plot, ks plot, 
raw residual and pearson residual.
}
