% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagpp.R
\name{diagpp}
\alias{diagpp}
\alias{diagpp.default}
\alias{diagpp.hp}
\alias{diagpp.mmhp}
\title{Generate diagnostic tools for different models, including quantile-quantile plot, ks plot, raw residual and pearson residual.}
\usage{
diagpp(
  object,
  t,
  pzt = NULL,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{diagpp}{default}(
  object,
  t,
  pzt = NULL,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{diagpp}{hp}(
  object,
  t,
  pzt = NULL,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{diagpp}{mmhp}(
  object,
  t,
  pzt = NULL,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)
}
\arguments{
\item{object}{a social network model}

\item{t}{event times}

\item{pzt}{probability for producing compensator}

\item{time.vec}{time segment to calculate the intensity for `numeric` method}

\item{latent.vec}{the probability of the latent space being in the active state}

\item{latent_event}{the estimated latent space of the point process model}
}
\value{
print qq plot and ks plot, and print out pearson and raw residuals.
}
\description{
Generate diagnostic tools for different models, including quantile-quantile plot, ks plot, raw residual and pearson residual.
}
