% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHPIntensity.R
\name{drawHPIntensity}
\alias{drawHPIntensity}
\title{Draw the intensity of Hawkes Process}
\usage{
drawHPIntensity(
  hp_obj,
  start,
  end,
  history = 0,
  events,
  color = 1,
  i = 1,
  add = FALSE,
  fit = FALSE,
  plot_events = FALSE,
  vec = NULL,
  int_title = "Hawkes Intensity"
)
}
\arguments{
\item{hp_obj}{object parameters for Hawkes process}

\item{start}{the start time of current state}

\item{end}{the end time of current state}

\item{history}{the past event times}

\item{events}{the event times happened in this state}

\item{color}{A specification for the default plotting color.}

\item{i}{state number this corresponds to a state jump directly before
which is only important when using mmhp}

\item{add}{whether to add the hawkes intensity to an existing plot}

\item{fit}{a boolean indicating whether to fit a hp or 
use the passed object}

\item{plot_events}{a boolean indicating whether events
inputted will be plotted}

\item{vec}{vector of initial values of parameters used in fithp}

\item{int_title}{title of the intensity plot}
}
\description{
Draw the intensity of Hawkes Process, a helper function for
 'drawUniMMHPIntensity'
while also available independently
}
\examples{
\dontrun{
hp_obj <- hp(lambda0 = 0.1, alpha = 0.45, beta = 0.5)
sims <- simulatehp(hp_obj, start = 0, end = 20, history = 0)
events <- sims$t
drawHPIntensity(hp_obj, start = 0, end = max(events), history = 0, events)
}
}
