% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHPIntensity.R
\name{drawHPIntensity}
\alias{drawHPIntensity}
\title{Draw the intensity of Hawkes Process}
\usage{
drawHPIntensity(
  lambda0,
  i,
  alpha,
  beta,
  start,
  end,
  history,
  hawkes_time,
  color = 1,
  add = FALSE
)
}
\arguments{
\item{lambda0}{parameters for Hawkes process}

\item{i}{state number}

\item{alpha}{parameters for Hawkes process}

\item{beta}{parameters for Hawkes process}

\item{start}{the start time of current state}

\item{end}{the end time of current state}

\item{history}{the past event times}

\item{hawkes_time}{the event times happened in this state}

\item{color}{A specification for the default plotting color.}

\item{add}{whether to add the hawkes intensity to an existing plot}
}
\description{
Draw the intensity of Hawkes Process, a helper function for 'drawUniMMHPIntensity'
}
