% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawHPPIntensity.R
\name{drawHPPIntensity}
\alias{drawHPPIntensity}
\title{Draw Intensity of Homogeneous Poisson Process}
\usage{
drawHPPIntensity(hpp, color = "red")
}
\arguments{
\item{hpp}{object for homogeneous poisson process}

\item{color}{a specification for the default plotting color.}
}
\description{
Draw the intensity for homogeneous poisson process events
}
