% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithp.R
\name{fithp}
\alias{fithp}
\title{Determine the MLE of hawkes process numerically}
\usage{
fithp(vec, t, termination)
}
\arguments{
\item{vec}{vector of initial values}

\item{t}{event times}

\item{termination}{end of observation period}
}
\value{
a hp object indicating the maximum likelihood parameter values (lambda0,alpha,beta) for hawkes process.
}
\description{
Determine the MLE of hawkes process numerically
}
\examples{
init=rep(0.1,3)
hp_obj <- hp(lambda0 = 0.1,alpha = 0.45,beta = 0.5)
sims <- simulatehp(hp_obj,start = 0, end = 100, history = 0)
fithp(init,sims$t,max(sims$t))                  
}
