% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fithpp.R
\name{fithpp}
\alias{fithpp}
\title{Fit a homogeneous poisson process}
\usage{
fithpp(t)
}
\arguments{
\item{t}{vector containing the event times.}
}
\value{
an hpp object fitted
}
\description{
Compute maximum likelihood estimator for the hpp model with time events
}
\examples{
pois_y <- hpp(lambda = 1, end = 10)
events <- simulatehpp(pois_y)
fithpp(events)
}
