% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hawkes_max_intensity.R
\name{hawkes_max_intensity}
\alias{hawkes_max_intensity}
\title{max intensity of a hawkes process}
\usage{
hawkes_max_intensity(object, events)
}
\arguments{
\item{object}{parameters for Hawkes process.}

\item{events}{parameters for Hawkes process.}
}
\value{
max of intensity
}
\description{
max intensity of a hawkes process
}
