% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp.R
\name{hpp}
\alias{hpp}
\title{Create a homogeneous poisson process model}
\usage{
hpp(lambda, events = NULL)
}
\arguments{
\item{lambda}{rate of the poisson process}

\item{events}{event times}
}
\value{
hpp object
}
\description{
Create a homogeneous poisson process according to given parameters:
 lambda, and events.
If events is missing, then it means that data will be
 added later(e.g. simulated)
}
\examples{
hpp(lambda = 1)
}
