% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpp.R
\name{hpp}
\alias{hpp}
\title{Create a homogeneous poisson process model}
\usage{
hpp(lambda, start = 0, end = 1, n = NULL)
}
\arguments{
\item{lambda}{rate of the poisson process}

\item{start}{start time of events}

\item{end}{end time of events}

\item{n}{number of events}
}
\value{
hpp object
}
\description{
Create a homogeneous poisson process according to given parameters: lambda, start, end, and n.
}
