% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\alias{intensity.default}
\alias{intensity.mmhp}
\alias{intensity.hp}
\alias{intensity.mmpp}
\alias{intensity.hpp}
\title{Compute the intensity of a temporal point process}
\usage{
intensity(object, event, method = "numeric")

\method{intensity}{default}(object, event, method = "numeric")

\method{intensity}{mmhp}(object, event, method = "numeric")

\method{intensity}{hp}(object, event, method = "numeric")

\method{intensity}{mmpp}(object, event, method = "numeric")

\method{intensity}{hpp}(object, event, method = "numeric")
}
\arguments{
\item{object}{an object of MMHP/HP/HPP/MMPP}

\item{event}{the observed/simulated events, including event times and 
start/end of observation period}

\item{method}{the method used to calculate intensity.
The candidates are: `numeric`, `atevent`, and `integral`, default to `numeric`.}
}
\value{
The intensity function of MMHP/HP/HPP/MMPP
}
\description{
Take a point process object and compute the intensity 
function accordingly, either numerically or at each event.
For processes with a latent state this will be inferred.
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- pp_simulate(x, n = 10)
z <- intensity(x, y) 

}
