% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityqqplot.R
\name{intensityqqplot}
\alias{intensityqqplot}
\alias{intensityqqplot.default}
\alias{intensityqqplot.hp}
\alias{intensityqqplot.hpp}
\alias{intensityqqplot.mmhp}
\title{Draw the intensity and q-q plot}
\usage{
intensityqqplot(object, events, pzt, ...)

\method{intensityqqplot}{default}(object, events, pzt, ...)

\method{intensityqqplot}{hp}(object, events, pzt = NULL, ...)

\method{intensityqqplot}{hpp}(object, events, pzt, ...)

\method{intensityqqplot}{mmhp}(object, events, pzt, ...)
}
\arguments{
\item{object}{parameters for the models: hp, hpp, and mmhp}

\item{events}{event times}

\item{pzt}{probability for calculating rescaled-interevent times.
which is only important when using mmhp}

\item{...}{further arguments}
}
\description{
Draw the intensity and q-q plot for models
}
