% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityqqplot.R
\name{intensityqqplot}
\alias{intensityqqplot}
\alias{intensityqqplot.default}
\alias{intensityqqplot.hp}
\alias{intensityqqplot.hpp}
\alias{intensityqqplot.mmhp}
\title{Draw the intensity and q-q plot}
\usage{
intensityqqplot(object, start, end, history, color, i, events, pzt)

\method{intensityqqplot}{default}(object, start, end, history, color, i, events, pzt)

\method{intensityqqplot}{hp}(
  object,
  start,
  end,
  history,
  color = 1,
  i = 1,
  events,
  pzt = NULL
)

\method{intensityqqplot}{hpp}(
  object,
  start,
  end,
  history,
  color = 1,
  i = 1,
  events,
  pzt = NULL
)

\method{intensityqqplot}{mmhp}(object, start, end, history, color = 1, i = 1, events, pzt)
}
\arguments{
\item{object}{parameters for the models: hp, hpp, and mmhp}

\item{start}{the start time of current state}

\item{end}{the end time of current state}

\item{history}{the past event times}

\item{color}{A specification for the default plotting color for intensity plot.}

\item{i}{state number this corresponds to a state jump directly before}

\item{events}{event times}

\item{pzt}{probability for calculating rescaled-interevent times.
which is only important when using mmhp}
}
\description{
Draw the intensity and q-q plot for models
}
