% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityqqplot.R
\name{intensityqqplot}
\alias{intensityqqplot}
\alias{intensityqqplot.default}
\alias{intensityqqplot.hp}
\alias{intensityqqplot.hpp}
\alias{intensityqqplot.mmhp}
\title{Draw Intensity of Fitted Point Process and QQ-Plot of Rescaled Events}
\usage{
intensityqqplot(object, events, ...)

\method{intensityqqplot}{default}(object, events, ...)

\method{intensityqqplot}{hp}(object, events, ...)

\method{intensityqqplot}{hpp}(object, events, ...)

\method{intensityqqplot}{mmhp}(object, events, ...)
}
\arguments{
\item{object}{parameters for the models: hp, hpp, and mmhp}

\item{events}{event times}

\item{...}{further arguments
which is only important when using mmhp}
}
\description{
Draw the intensity and q-q plot for models
}
