% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglik.R
\name{negloglik}
\alias{negloglik}
\title{Compute negative log likelihood for point process models}
\usage{
negloglik(object, events, end)
}
\arguments{
\item{object}{point process object containing the parameters}

\item{events}{vector containing the event times.}

\item{end}{the end time of event times}
}
\value{
a scalar indicating the negative log likelihood
}
\description{
Compute negative log likelihood for point process models
 with model specified time events or simulated time events
}
\keyword{internal}
