% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonresidual.R
\name{pearsonresidual}
\alias{pearsonresidual}
\title{Compute Pearson residuals for point process models}
\usage{
pearsonresidual(object, events, start, end, steps = 1000)
}
\arguments{
\item{object}{social network model containing the parameters}

\item{events}{vector of event times}

\item{start}{start of observation period (default 0)}

\item{end}{termination time (default final event)}

\item{steps}{number of steps for numeric integration (if needed)}
}
\value{
the Pearson residual
}
\description{
Compute Pearson residuals for point processes 
with specified parameters and events.
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- pp_mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- pp_simulate(x, n = 10)
pearsonresidual(x, events = y$events[-1])
}
