% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonresidual.R
\name{pearsonresidual}
\alias{pearsonresidual}
\alias{pearsonresidual.default}
\alias{pearsonresidual.mmhp}
\alias{pearsonresidual.hp}
\alias{pearsonresidual.hpp}
\title{Compute Pearson residuals for social network models}
\usage{
pearsonresidual(object, events, start, end)

\method{pearsonresidual}{default}(object, events, start = min(events), end = max(events))

\method{pearsonresidual}{mmhp}(object, events, start = min(events), end = max(events))

\method{pearsonresidual}{hp}(object, events, start = min(events), end = max(events))

\method{pearsonresidual}{hpp}(object, events, start = min(events), end = max(events))
}
\arguments{
\item{object}{social network model containing the parameters}

\item{events}{vector of event happening time}

\item{start}{start of observation period}

\item{end}{termination time}
}
\value{
the Pearson residual
}
\description{
Compute Pearson residuals for social network models 
with model specified time events or simulated time events
}
