% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkPearsonresidual.R, R/pearsonresidual.R
\name{pearsonresidual}
\alias{pearsonresidual}
\alias{pearsonresidual.default}
\alias{networkPearsonresidual.hp}
\alias{pearsonresidual.mmhp}
\alias{pearsonresidual.hp}
\title{Compute network based Pearson residuals for social network models}
\usage{
pearsonresidual(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{pearsonresidual}{default}(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

networkPearsonresidual.hp(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

pearsonresidual(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{pearsonresidual}{default}(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{pearsonresidual}{mmhp}(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)

\method{pearsonresidual}{hp}(
  object,
  t,
  termination,
  time.vec = NULL,
  latent.vec = NULL,
  latent_event = NULL
)
}
\arguments{
\item{object}{social network model containing the parameters}

\item{t}{vector of event happening time}

\item{termination}{termination time}

\item{time.vec}{time segment to calculate the intensity for `numeric` method}

\item{latent.vec}{the probability of the latent space being in the active state}

\item{latent_event}{the estimated latent space of the point process model}
}
\value{
a dataframe of three columns: Send, Rec and the corresponsding Pearson residual

the pearson residual
}
\description{
Compute network based Pearson residuals for social network models with model specified time events or simulated time events

Compute Pearson residuals for social network models with model specified time events or simulated time events
}
