% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothpp.R
\name{plothpp_step}
\alias{plothpp_step}
\title{Draw Homogeneous Poisson Process}
\usage{
plothpp_step(events, color = "red")
}
\arguments{
\item{events}{homogeneous poisson process events}

\item{color}{a specification for the default plotting color.}
}
\description{
Draw the event time counts for homogeneous poisson process
}
