% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_simulate.R
\name{pp_simulate}
\alias{pp_simulate}
\title{Simulate events from a temporal point process}
\usage{
pp_simulate(object, start = 0, end = 1, n = NULL)
}
\arguments{
\item{object}{point process model object of type hpp, hp, or mmhp}

\item{start}{start time of events simulated}

\item{end}{end time of events simulated}

\item{n}{number of events}
}
\value{
a vector of event times
}
\description{
Currently available point processes are homogeneous Poisson,
Hawkes with exponential kernel and MMHP
}
\examples{
hpp_obj <- hpp(lambda = 1)
s <- pp_simulate(hpp_obj, end = 10, n=50)
}
