% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawresidual.R
\name{rawresidual}
\alias{rawresidual}
\alias{rawresidual.default}
\alias{rawresidual.hp}
\alias{rawresidual.mmhp}
\alias{rawresidual.hpp}
\title{Compute raw residuals for social network models}
\usage{
rawresidual(object, events, termination)

\method{rawresidual}{default}(object, events, termination)

\method{rawresidual}{hp}(object, events, termination)

\method{rawresidual}{mmhp}(object, events, termination)

\method{rawresidual}{hpp}(object, events, termination)
}
\arguments{
\item{object}{social network model contating the parameters}

\item{events}{vector of event happening time}

\item{termination}{termination time}
}
\value{
the raw residual
}
\description{
Compute raw residuals for social network models with model specified time events or simulated time events
}
