% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawresidual.R
\name{rawresidual}
\alias{rawresidual}
\alias{rawresidual.default}
\alias{rawresidual.hp}
\alias{rawresidual.mmhp}
\alias{rawresidual.hpp}
\title{Compute raw residuals for point process models}
\usage{
rawresidual(object, events, start, end)

\method{rawresidual}{default}(object, events, start = 0, end = max(events))

\method{rawresidual}{hp}(object, events, start = 0, end = max(events))

\method{rawresidual}{mmhp}(object, events, start = 0, end = max(events))

\method{rawresidual}{hpp}(object, events, start = 0, end = max(events))
}
\arguments{
\item{object}{point process model containing the parameters}

\item{events}{vector of event times}

\item{start}{start of observation period (default 0)}

\item{end}{end of observation period (default final event)}
}
\value{
the raw residual
}
\description{
Compute raw residuals for for point processes 
with specified parameters and events.
}
\examples{
Q <- matrix(c(-0.4, 0.4, 0.2, -0.2), ncol = 2, byrow = TRUE)
x <- mmhp(Q, delta = c(1 / 3, 2 / 3), lambda0 = 0.9, 
lambda1 = 1.1, alpha = 0.8, beta = 1.2)
y <- pp_simulate(x, n = 10)
rawresidual(x, events = y$events[-1])
}
