% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehp.R
\name{simulatehp}
\alias{simulatehp}
\title{Simulate Hawkes process during active state}
\usage{
simulatehp(lambda0, alpha, beta, start, end, history)
}
\arguments{
\item{lambda0}{parameters for Hawkes process.}

\item{alpha}{parameters for Hawkes process.}

\item{beta}{parameters for Hawkes process.}

\item{start}{start time of the Hawkes process.}

\item{end}{end time of the Hawkes process.}

\item{history}{the past event times.}
}
\value{
simulated Hawkes Process
}
\description{
Simulate Hawkes process during active state (including all the histoty), a helper function for "simulationmmhp"
}
