% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehp.R
\name{simulatehp}
\alias{simulatehp}
\title{Simulate Hawkes process during active state}
\usage{
simulatehp(hp, start = 0, end, history = 0)
}
\arguments{
\item{hp}{hawkes process object, including parameters in list type 
(lambda0, alpha, beta, tau)}

\item{start}{start time of the Hawkes process.}

\item{end}{end time of the Hawkes process.}

\item{history}{the past event times.}
}
\value{
simulated Hawkes Process
}
\description{
Simulate Hawkes process during active state (including all the histoty),
 a helper function for "simulationmmhp"
}
\examples{
hp_obj <- hp(lambda0 = 0.1,alpha = 0.45,beta = 0.5)
simulatehp(hp_obj,start = 0, end = 100)
}
