% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehpp.R
\name{simulatehpp}
\alias{simulatehpp}
\title{Simulate homogeneous poisson process events}
\usage{
simulatehpp(lambda, end = 1, start = 0, n = NULL)
}
\arguments{
\item{lambda}{rate of the poisson process}

\item{end}{end time of events}

\item{start}{start time of events}

\item{n}{number of events}
}
\value{
a vector of length n
}
\description{
Simulate homogeneous poisson process events
}
