% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehpp.R
\name{simulatehpp}
\alias{simulatehpp}
\title{Simulate homogeneous poisson process events}
\usage{
simulatehpp(hpp, start = 0, end = NULL, n = NULL, seed = NULL)
}
\arguments{
\item{hpp}{hpp object in list type}

\item{start}{start time of events simulated}

\item{end}{end time of events simulated}

\item{n}{number of events}

\item{seed}{seed for simulation}
}
\value{
a vector of length n
}
\description{
Simulate homogeneous poisson process events
}
\examples{
hpp_obj=hpp(lambda = 1)
s=simulatehpp(hpp_obj, end = 10, n=50)
hist(s)
}
