% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehpp.R
\name{simulatehpp}
\alias{simulatehpp}
\title{Simulate homogeneous poisson process events}
\usage{
simulatehpp(hpp)
}
\arguments{
\item{hpp}{hpp object in list type, (lambda, start=0, end=1, n = NULL)}
}
\value{
a vector of length n
}
\description{
Simulate homogeneous poisson process events
}
